--- src/components/Menu/index.js
@@ -1,6 +1,0 @@
-/**
- * Created by yangyang on 2017/9/13.
- */
-import ComposeMenu from './ComposeMenu'
-
-export default ComposeMenu
