+++ src/routes/Home/ContentRouter.js
@@ -3,9 +3,8 @@
  */
 import React from 'react'
 import {Route, Switch} from 'react-router-dom'
-import AuthRoute from '../../routes/AuthRoute'
-import Dashboard from '../../components/Dashboard/Dashboard'
-import SysConfig from '../../components/System/SysConfig'
+import Dashboard from '../../components/Dashboard'
+import SysConfig from '../../components/SysConfig'
 
 const ContentRouter = ({match}) => {
   return (
